+++ app/utils/shortnameToUnicode/ascii.js
@@ -4,4 +4,7 @@
 /* eslint-disable comma-spacing */
 /* eslint-disable key-spacing */
 const ascii = {'*\\0/*':{unicode:'🙆'},'*\\O/*':{unicode:'🙆'},'-___-':{unicode:'😑'},':\'-)':{unicode:'😂'},'\':-)':{unicode:'😅'},'\':-D':{unicode:'😅'},'>:-)':{unicode:'😆'},'\':-(':{unicode:'😓'},'>:-(':{unicode:'😠'},':\'-(':{unicode:'😢'},'O:-)':{unicode:'😇'},'0:-3':{unicode:'😇'},'0:-)':{unicode:'😇'},'0;^)':{unicode:'😇'},'O;-)':{unicode:'😇'},'0;-)':{unicode:'😇'},'O:-3':{unicode:'😇'},'-__-':{unicode:'😑'},':-Þ':{unicode:'😛'},'</3':{unicode:'💔'},':\')':{unicode:'😂'},':-D':{unicode:'😃'},'\':)':{unicode:'😅'},'\'=)':{unicode:'😅'},'\':D':{unicode:'😅'},'\'=D':{unicode:'😅'},'>:)':{unicode:'😆'},'>;)':{unicode:'😆'},'>=)':{unicode:'😆'},';-)':{unicode:'😉'},'*-)':{unicode:'😉'},';-]':{unicode:'😉'},';^)':{unicode:'😉'},'\':(':{unicode:'😓'},'\'=(':{unicode:'😓'},':-*':{unicode:'😘'},':^*':{unicode:'😘'},'>:P':{unicode:'😜'},'X-P':{unicode:'😜'},'>:[':{unicode:'😞'},':-(':{unicode:'😞'},':-[':{unicode:'😞'},'>:(':{unicode:'😠'},':\'(':{unicode:'😢'},';-(':{unicode:'😢'},'>.<':{unicode:'😣'},'#-)':{unicode:'😵'},'%-)':{unicode:'😵'},'X-)':{unicode:'😵'},'\\0/':{unicode:'🙆'},'\\O/':{unicode:'🙆'},'0:3':{unicode:'😇'},'0:)':{unicode:'😇'},'O:)':{unicode:'😇'},'O=)':{unicode:'😇'},'O:3':{unicode:'😇'},'B-)':{unicode:'😎'},'8-)':{unicode:'😎'},'B-D':{unicode:'😎'},'8-D':{unicode:'😎'},'-_-':{unicode:'😑'},'>:\\':{unicode:'😕'},'>:/':{unicode:'😕'},':-/':{unicode:'😕'},':-.':{unicode:'😕'},':-P':{unicode:'😛'},':Þ':{unicode:'😛'},':-b':{unicode:'😛'},':-O':{unicode:'😮'},'O_O':{unicode:'😮'},'>:O':{unicode:'😮'},':-X':{unicode:'😶'},':-#':{unicode:'😶'},':-)':{unicode:'🙂'},'(y)':{unicode:'👍'},'<3':{unicode:'❤'},'=D':{unicode:'😃'},';)':{unicode:'😉'},'*)':{unicode:'😉'},';]':{unicode:'😉'},';D':{unicode:'😉'},':*':{unicode:'😘'},'=*':{unicode:'😘'},':(':{unicode:'😞'},':[':{unicode:'😞'},'=(':{unicode:'😞'},':@':{unicode:'😠'},';(':{unicode:'😢'},'D:':{unicode:'😨'},':$':{unicode:'😳'},'=$':{unicode:'😳'},'#)':{unicode:'😵'},'%)':{unicode:'😵'},'X)':{unicode:'😵'},'B)':{unicode:'😎'},'8)':{unicode:'😎'},':/':{unicode:'😕'},':\\':{unicode:'😕'},'=/':{unicode:'😕'},'=\\':{unicode:'😕'},':L':{unicode:'😕'},'=L':{unicode:'😕'},':P':{unicode:'😛'},'=P':{unicode:'😛'},':b':{unicode:'😛'},':O':{unicode:'😮'},':X':{unicode:'😶'},':#':{unicode:'😶'},'=X':{unicode:'😶'},'=#':{unicode:'😶'},':)':{unicode:'🙂'},'=]':{unicode:'🙂'},'=)':{unicode:'🙂'},':]':{unicode:'🙂'},':D':{unicode:'😄'}};
+
+export const asciiRegexp = '(\\*\\\\0\\/\\*|\\*\\\\O\\/\\*|\\-___\\-|\\:\'\\-\\)|\'\\:\\-\\)|\'\\:\\-D|\\>\\:\\-\\)|>\\:\\-\\)|\'\\:\\-\\(|\\>\\:\\-\\(|>\\:\\-\\(|\\:\'\\-\\(|O\\:\\-\\)|0\\:\\-3|0\\:\\-\\)|0;\\^\\)|O;\\-\\)|0;\\-\\)|O\\:\\-3|\\-__\\-|\\:\\-Þ|\\:\\-Þ|\\<\\/3|<\\/3|\\:\'\\)|\\:\\-D|\'\\:\\)|\'\\=\\)|\'\\:D|\'\\=D|\\>\\:\\)|>\\:\\)|\\>;\\)|>;\\)|\\>\\=\\)|>\\=\\)|;\\-\\)|\\*\\-\\)|;\\-\\]|;\\^\\)|\'\\:\\(|\'\\=\\(|\\:\\-\\*|\\:\\^\\*|\\>\\:P|>\\:P|X\\-P|\\>\\:\\[|>\\:\\[|\\:\\-\\(|\\:\\-\\[|\\>\\:\\(|>\\:\\(|\\:\'\\(|;\\-\\(|\\>\\.\\<|>\\.<|#\\-\\)|%\\-\\)|X\\-\\)|\\\\0\\/|\\\\O\\/|0\\:3|0\\:\\)|O\\:\\)|O\\=\\)|O\\:3|B\\-\\)|8\\-\\)|B\\-D|8\\-D|\\-_\\-|\\>\\:\\\\|>\\:\\\\|\\>\\:\\/|>\\:\\/|\\:\\-\\/|\\:\\-\\.|\\:\\-P|\\:Þ|\\:Þ|\\:\\-b|\\:\\-O|O_O|\\>\\:O|>\\:O|\\:\\-X|\\:\\-#|\\:\\-\\)|\\(y\\)|\\<3|<3|\\=D|;\\)|\\*\\)|;\\]|;D|\\:\\*|\\=\\*|\\:\\(|\\:\\[|\\=\\(|\\:@|;\\(|D\\:|\\:\\$|\\=\\$|#\\)|%\\)|X\\)|B\\)|8\\)|\\:\\/|\\:\\\\|\\=\\/|\\=\\\\|\\:L|\\=L|\\:P|\\=P|\\:b|\\:O|\\:X|\\:#|\\=X|\\=#|\\:\\)|\\=\\]|\\=\\)|\\:\\]|\\:D)';
+
 export default ascii;
